Update dropdown to properly close on click outside of container
Before, the nav dropdown had it's own div that would be sized to 100% of
the screen height and width. This would be an issue, particularly in
chrome (as far as I tested) because it would add a second scrollbar that
would interfere with the normal window scrollbar. With the new solution,
the user can close the nav dropdown when clicking outside of the
dropdown component; the user can now also scroll the entire window without issue.